// .agent/skills/umbraco-backoffice/examples/tree-example/Client/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type OurTreeItemResponseModel = {
    hasChildren: boolean;
    id: string;
    parent?: ReferenceByIdModel | null;
    name: string;
    icon: string;
};

export type PagedOurTreeItemResponseModel = {
    total: number;
    items: Array<OurTreeItemResponseModel>;
};

export type ReferenceByIdModel = {
    id: string;
};

export type GetAncestorsData = {
    body?: never;
    path?: never;
    query?: {
        id?: string;
    };
    url: '/umbraco/umbtreeclient/api/v1/Ancestors';
};

export type GetAncestorsErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
};

export type GetAncestorsResponses = {
    /**
     * OK
     */
    200: Array<OurTreeItemResponseModel>;
};

export type GetAncestorsResponse = GetAncestorsResponses[keyof GetAncestorsResponses];

export type GetChildrenData = {
    body?: never;
    path?: never;
    query?: {
        parent?: string;
        skip?: number;
        take?: number;
    };
    url: '/umbraco/umbtreeclient/api/v1/Children';
};

export type GetChildrenErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
};

export type GetChildrenResponses = {
    /**
     * OK
     */
    200: PagedOurTreeItemResponseModel;
};

export type GetChildrenResponse = GetChildrenResponses[keyof GetChildrenResponses];

export type PingData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/umbraco/umbtreeclient/api/v1/ping';
};

export type PingErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
};

export type PingResponses = {
    /**
     * OK
     */
    200: string;
};

export type PingResponse = PingResponses[keyof PingResponses];

export type GetRootData = {
    body?: never;
    path?: never;
    query?: {
        skip?: number;
        take?: number;
        foldersOnly?: boolean;
    };
    url: '/umbraco/umbtreeclient/api/v1/root';
};

export type GetRootErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
};

export type GetRootResponses = {
    /**
     * OK
     */
    200: PagedOurTreeItemResponseModel;
};

export type GetRootResponse = GetRootResponses[keyof GetRootResponses];

export type ClientOptions = {
    baseUrl: 'https://localhost:44348' | (string & {});
};